yeah

+++ core/U.c
@@ -41,13 +41,14 @@ static inline void sampleWithoutReplacement(const size_t populationSize, const s
 	}
 }
 
+/* 
 static unsigned long long int binomialCoefficient(size_t n, size_t k) {
 	if (k == 0) return 1;
 	if (k == 1) return n;
 	if (k > n / 2) return binomialCoefficient(n, n - k);
 	return (int)n * binomialCoefficient(n - 1, k - 1) / (int)k;
 }
-
+*/
 
 typedef struct {
 	unsigned int i;
@@ -116,8 +117,8 @@ double U(
 		gsl_combination* cmb = gsl_combination_calloc(n, m);
 		int b = 0;
 		do {
-			for (int i = 0; i < m; i++) {
-				for (int j = 0; j < d; j++) gsl_matrix_set(subsample, i, j, gsl_matrix_get(data, gsl_combination_data(cmb)[i], j));
+		  for (size_t i = 0; i < (unsigned int) m; i++) {
+				for (size_t j = 0; j < d; j++) gsl_matrix_set(subsample, i, j, gsl_matrix_get(data, gsl_combination_data(cmb)[i], j));
 			}
 			double newval = kernel(subsample);
 			gsl_vector_set(resamplingResults, b++, newval);
@@ -130,8 +131,8 @@ double U(
 		size_t * indices = malloc(m * sizeof(size_t));
 		for (size_t b = 0; b < B; b++) {
 			sampleWithoutReplacement(n, m, indices, r);
-			for (int i = 0; i < m; i++) {
-				for (int j = 0; j < d; j++) gsl_matrix_set(subsample, i, j, gsl_matrix_get(data, indices[i], j));
+			for (size_t i = 0; i < (unsigned int) m; i++) {
+			  for (size_t j = 0; j < (unsigned int) d; j++) gsl_matrix_set(subsample, i, j, gsl_matrix_get(data, indices[i], j));
 			}
 			double newval = kernel(subsample);
 			gsl_vector_set(resamplingResults, b, newval);
